requirements.txt
asgiref==3.7.2
coverage==7.4.3
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
parameterized==0.9.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
pre-commit==3.6.2
flake8==7.0.0
black==24.2.0
